requirements.txt
Django~=5.0.1
gunicorn~=21.2.0
Pillow~=10.2.0
djangorestframework